+++ PFGAP/src/core/AppContextSnapshot.java
@@ -1,7 +1,7 @@
 package core;
 
 import distance.MEASURE;
-import imputation.Imputer;
+
 import java.io.Serializable;
 import java.util.*;
