+++ PFGAP/src/imputation/initial/GlobalMedianImpute.java
@@ -1,6 +1,8 @@
-package imputation;
+package imputation.initial;
 
 import datasets.ListObjectDataset;
+import imputation.util.MissingIndices;
+
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
