+++ PFGAP/src/imputation/initial/GlobalModeImpute.java
@@ -1,4 +1,4 @@
-package imputation;
+package imputation.initial;
 
 import datasets.ListObjectDataset;
 import java.util.*;
