+++ PFGAP/src/imputation/initial/Imputer.java
@@ -1,4 +1,4 @@
-package imputation;
+package imputation.initial;
 
 import datasets.ListObjectDataset;
