+++ PFGAP/src/imputation/initial/MaskedDistance.java
@@ -1,4 +1,4 @@
-package imputation;
+package imputation.initial;
 
 import java.util.*;
