+++ PFGAP/src/imputation/initial/MedianImpute.java
@@ -1,16 +1,13 @@
-package imputation;
+package imputation.initial;
 
-import core.AppContext;
 import datasets.ListObjectDataset;
-import util.Statistics;
+import imputation.util.MissingIndices;
 
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-// We need to be working with List<Double[]> or List<Double[][]>
-
-public class MeanImpute extends Imputer {
+public class MedianImpute extends Imputer {
 
     private static final int CHUNK_SIZE = 1000;
 
@@ -23,7 +20,6 @@ public class MeanImpute extends Imputer {
 
             for (int i = 0; i < rawData.size(); i += CHUNK_SIZE) {
                 int end = Math.min(i + CHUNK_SIZE, rawData.size());
-
                 List<Object> chunk = rawData.subList(i, end);
                 List<List<List<Integer>>> missingChunk = mi.indices2D.subList(i, end);
 
@@ -38,23 +34,16 @@ public class MeanImpute extends Imputer {
                         .collect(Collectors.toList());
 
                 result.addAll(processedChunk);
-
-                // Erase original data
-                for (int j = i; j < end; j++) {
-                    rawData.set(j, null);
-                }
-
-                System.gc(); // Hint GC
+                for (int j = i; j < end; j++) rawData.set(j, null);
+                System.gc();
             }
 
             missingDS.setData(result);
-
         } else {
             List<Object> result = new ArrayList<>(rawData.size());
 
             for (int i = 0; i < rawData.size(); i += CHUNK_SIZE) {
                 int end = Math.min(i + CHUNK_SIZE, rawData.size());
-
                 List<Object> chunk = rawData.subList(i, end);
                 List<List<Integer>> missingChunk = mi.indices1D.subList(i, end);
 
@@ -69,13 +58,8 @@ public class MeanImpute extends Imputer {
                         .collect(Collectors.toList());
 
                 result.addAll(processedChunk);
-
-                // Erase original data
-                for (int j = i; j < end; j++) {
-                    rawData.set(j, null);
-                }
-
-                System.gc(); // Hint GC
+                for (int j = i; j < end; j++) rawData.set(j, null);
+                System.gc();
             }
 
             missingDS.setData(result);
@@ -84,26 +68,19 @@ public class MeanImpute extends Imputer {
 
     public static double[] convert1DPrimitive(Double[] row, List<Integer> missingIndices) {
         double[] result = new double[row.length];
-        double sum = 0;
-        int count = 0;
-
-        Set<Integer> missingSet = new HashSet<>(missingIndices);
+        List<Double> values = new ArrayList<>();
 
         for (int i = 0; i < row.length; i++) {
-            if (!missingSet.contains(i)) {
-                Double val = row[i];
-                if (val != null) {
-                    sum += val;
-                    count++;
-                    result[i] = val;
-                }
+            if (row[i] != null) {
+                result[i] = row[i];
+                values.add(row[i]);
             }
         }
 
-        double mean = count > 0 ? sum / count : 0;
+        double median = computeMedian(values);
 
         for (int i : missingIndices) {
-            result[i] = mean;
+            result[i] = median;
         }
 
         return result;
@@ -111,72 +88,20 @@ public class MeanImpute extends Imputer {
 
     public static double[][] convert2DPrimitive(Double[][] matrix, List<List<Integer>> missingIndices) {
         double[][] result = new double[matrix.length][];
-
         for (int i = 0; i < matrix.length; i++) {
             Double[] row = matrix[i];
             List<Integer> missing = missingIndices.get(i);
             result[i] = convert1DPrimitive(row, missing);
         }
-
         return result;
+    }
+
+    private static double computeMedian(List<Double> values) {
+        if (values.isEmpty()) return 0.0;
+        Collections.sort(values);
+        int n = values.size();
+        return (n % 2 == 1)
+                ? values.get(n / 2)
+                : (values.get(n / 2 - 1) + values.get(n / 2)) / 2.0;
     }
 }
-
-
-/*public class ListMeanImpute extends Imputer{
-
-
-    public static void Impute(ListObjectDataset missingDS){
-        List<Object> rawData = missingDS.getData();
-        missingDS.setData(rawData);
-    }
-
-    // assumes we would want to work with ListDataset--given Double[], need double[].
-
-    public static double[] convertToPrimitive(Object Input) {
-        Double[] input = (Double[]) Input;
-        int len = input.length;
-        double[] result = new double[len];
-        double sum = 0;
-        int count = 0;
-
-        for (int i = 0; i < len; i++) {
-            Double val = input[i];
-            if (val != null) {
-                sum += val;
-                count++;
-                result[i] = val;
-            }
-        }
-
-        double mean = count > 0 ? sum / count : 0;
-
-        for (int i = 0; i < len; i++) {
-            if (input[i] == null) {
-                result[i] = mean;
-            }
-        }
-
-        return result;
-    }
-
-
-
-    public static List<double[]> convertList(List<Object> inputList) {
-        return inputList.parallelStream()
-                .map(ListMeanImpute::convertToPrimitive)
-                .collect(Collectors.toList());
-    }
-
-
-
-    // this is the non parallel version:
-    //public static List<double[]> convertList(List<Double[]> inputList) {
-    //    List<double[]> resultList = new ArrayList<>();
-    //    for (Double[] row : inputList) {
-    //        resultList.add(convertToPrimitive(row));
-    //    }
-    //    return resultList;
-    //}
-
-}*/
