+++ PFGAP/src/imputation/util/MissingIndices.java
@@ -1,4 +1,4 @@
-package imputation;
+package imputation.util;
 
 import java.io.Serializable;
 import java.util.List;
