+++ PFGAP/src/imputation/util/MissingIndicesBuilder.java
@@ -1,6 +1,4 @@
-package imputation;
-
-import core.AppContext;
+package imputation.util;
 
 import java.util.ArrayList;
 import java.util.List;
